[chore] CI/CD settings (#13)
* chore: common 폴더 구조 변경

* chore: 환경변수 파일 분리 (local/dev)

* feat: 서버 profile 확인용 api 추가

* chore: 무중단 배포시 실행할 스크립트 작성

* chore: Github Actions 스크립트 작성 (임시 trigger 설정)

* chore: CI/CD trigger 재설정

* chore: 인덴트 오류 수정

* chore: CodeDeploy 옵션 문자열 오류 수정

* chore: Spring Security 및 Redis 임시 주석 처리

* chore: trigger 시점 develop 으로 변경

* chore: profile 조회 엔드포인트 수정

* style: CD 브랜치 trigger 시점 임시 변경

* fix: trigger 시점 develop 으로 원상 복귀

---------
