+++ scripts/appspec.yml
@@ -1,0 +1,22 @@
+version: 0.0
+os: linux
+
+files:
+  - source:  /
+    destination: /home/ubuntu/sanhak-server
+    overwrite: yes
+
+permissions:
+  - object: /
+    pattern: "**"
+    owner: ubuntu
+    group: ubuntu
+
+hooks:
+  AfterInstall:
+    - location: deploy.sh
+      timeout: 180
+      runas: ubuntu
+    - location: switch.sh
+      timeout: 180
+      runas: ubuntu
