+++ src/main/java/org/kau/kkoolbeeServer/global/common/controller/ServerProfileController.java
@@ -1,0 +1,22 @@
+package org.kau.kkoolbeeServer.global.common.controller;
+
+import lombok.RequiredArgsConstructor;
+import org.springframework.core.env.Environment;
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.RestController;
+
+import java.util.Arrays;
+
+@RestController
+@RequiredArgsConstructor
+public class ServerProfileController {
+
+    private final Environment env;
+
+    @GetMapping("/profile")
+    public String getProfile() {
+        return Arrays.stream(env.getActiveProfiles())
+                .findFirst()
+                .orElse("");
+    }
+}
