+++ src/main/java/org/kau/kkoolbeeServer/global/common/dto/enums/ErrorType.java
@@ -1,4 +1,4 @@
-package org.kau.kkoolbeeServer.global.common.enums;
+package org.kau.kkoolbeeServer.global.common.dto.enums;
 
 import lombok.AccessLevel;
 import lombok.AllArgsConstructor;
@@ -7,12 +7,17 @@ import org.springframework.http.HttpStatus;
 
 @Getter
 @AllArgsConstructor(access = AccessLevel.PRIVATE)
-public enum SuccessType {
+public enum ErrorType {
 
     /**
-     * 200 OK
+     * 400 BAD REQUEST
      */
-    PROCESS_SUCCESS(HttpStatus.OK, "OK"),
+    REQUEST_VALIDATION_EXCEPTION(HttpStatus.BAD_REQUEST, "잘못된 요청입니다"),
+
+    /**
+     * 500 INTERNAL SERVER ERROR
+     */
+    INTERNAL_SERVER_ERROR(HttpStatus.INTERNAL_SERVER_ERROR, "알 수 없는 서버 에러가 발생했습니다."),
     ;
 
     private final HttpStatus httpStatus;
