+++ src/main/resources/application-dev2.yml
@@ -1,0 +1,23 @@
+spring:
+  config:
+    activate:
+      on-profile: dev2
+
+  datasource:
+    driver-class-name: com.mysql.cj.jdbc.Driver
+    url: ${DB_URL_DEV}
+    username: ${DB_USER_DEV}
+    password: ${DB_PWD_DEV}
+  jpa:
+    show-sql: false
+    hibernate:
+      ddl-auto: update
+      ejb:
+        naming_strategy: org.springframework.boot.orm.jpa.hibernate.SpringPhysicalNamingStrategy
+    properties:
+      hibernate:
+        auto_quote_keyword: true
+        format_sql: true
+
+server:
+  port: 8082
